legacy/requirements_raspberry_pi_lite.txt
# AI桌宠树莓派轻量级依赖文件
# 适用于Python 3.7-3.9，去除了重型依赖

# Web框架
flask>=2.0.1,<3.0.0
jinja2>=3.0.0,<4.0.0
werkzeug>=2.0.0,<3.0.0
markupsafe>=2.0.0,<3.0.0
itsdangerous>=2.0.0,<3.0.0
click>=8.0.0,<9.0.0

# 图像处理
opencv-python-headless>=4.5.0,<5.0.0
numpy>=1.19.0,<2.0.0
Pillow>=9.0.0

# 树莓派硬件接口
RPi.GPIO>=0.7.0
picamera>=1.13

# 音频处理
SpeechRecognition>=3.8.1
pyaudio>=0.2.11
pocketsphinx>=0.1.15
pygame>=2.1.0

# AI和语音服务
google-generativeai>=0.3.0
pvporcupine>=3.0.0
edge-tts>=6.1.0
jieba>=0.42.1

# OLED显示
adafruit-circuitpython-ssd1306>=2.12.0
adafruit-blinka>=8.0.0

# 系统工具
requests>=2.25.0
aiofiles>=0.8.0

# 数据处理
# sqlite3 内置在Python中，无需安装